+++ demo/playground/src/jsMain/kotlin/playgroundMain.kt
@@ -1,5 +1,5 @@
 import space.kscience.dataforge.misc.DFExperimental
-import space.kscience.plotly.PlotlyPlugin
+import space.kscience.plotly.PlotlyJSPlugin
 import space.kscience.visionforge.html.runVisionClient
 import space.kscience.visionforge.jupyter.VFNotebookClient
 import space.kscience.visionforge.markup.MarkupPlugin
@@ -9,7 +9,7 @@ import space.kscience.visionforge.tables.TableVisionJsPlugin
 @DFExperimental
 fun main() = runVisionClient {
     plugin(ThreePlugin)
-    plugin(PlotlyPlugin)
+    plugin(PlotlyJSPlugin)
     plugin(MarkupPlugin)
     plugin(TableVisionJsPlugin)
     plugin(VFNotebookClient)