+++ demo/playground/src/jvmMain/kotlin/allThingsDemo.kt
@@ -2,6 +2,7 @@ package space.kscience.visionforge.examples
 
 import kotlinx.html.h2
 import space.kscience.dataforge.meta.ValueType
+import space.kscience.plotly.PlotlyPlugin
 import space.kscience.plotly.layout
 import space.kscience.plotly.models.ScatterMode
 import space.kscience.plotly.models.TextPosition
@@ -43,6 +44,7 @@ fun main() = makeVisionFile(
 
     h2 { +"Interactive plots with Plotly" }
     vision("plot") {
+        requirePlugin(PlotlyPlugin)
         plotly {
             scatter {
                 x(1, 2, 3, 4)