+++ visionforge-solid/src/commonMain/kotlin/space/kscience/visionforge/solid/LightSource.kt
@@ -17,7 +17,7 @@ public abstract class LightSource : MiscSolid() {
     override val descriptor: MetaDescriptor get() = LightSource.descriptor
 
     public val color: ColorAccessor by colorProperty(SolidMaterial.COLOR_KEY)
-    public var intensity: Number by properties.number(INTENSITY_KEY) { 1.0 }
+    public var intensity: Number by properties.number(INTENSITY_KEY) { 2.0 }
 
     public companion object {
         public val INTENSITY_KEY: Name = "intensity".asName()